Added new Utils structure
